split Types and Parser to their files
