+++ hob.cabal
@@ -38,6 +38,8 @@ executable hob
 library
     Exposed-modules:     Hob.Ui, Hob.Ui.Editor, Hob.Ui.Sidebar, Hob.Ui.CommandEntry,
                          Hob.Ui.SidebarSearch, Hob.Ui.Editor.Search, Hob.Ui.Editor.Fancy,
+                         Hob.Ui.Editor.Fancy.Types,
+                         Hob.Ui.Editor.Fancy.Parser,
                          Hob.Command.FindText, Hob.Command.ReplaceText, Hob.Command.CloseCurrentTab,
                          Hob.Command.SaveCurrentTab, Hob.Command.NewTab, Hob.Command.FocusCommandEntry,
                          Hob.Command.FocusNextTab, Hob.Command.ReloadSidebarTree, Hob.Command.FocusSidebar,
